[nativeio] NativeIO: Add setLength via workaround
The MacOS sandbox does not allow calling setLength from the renderer
directly, see crbug.com/1084565. The call to setLength is therefore
routed through the browser process, which likely decreases performance
and increases complexity. Another option would be to create a
SeatbeltExtensionToken that allows to pierce the sandbox on MacOS for a
single file. As this would likely lead to more complex code, we decided
that going through the browser process is the better option for now.

Bug: 914488
Change-Id: Idc58d804e19751a2e37cfeb5c645f778edf4639c