[LayoutNG] Move relative position calculation into NG.
This patch moves the relative position calculation into NG. Previously
we needed to manually account for the relative position of an object
in many different places, overflow calculation, geometry APIs etc.

This patch directly updates the child's offset (relative to its parent)
to directly incorporate this additional offset.

This calculation occurs within the NGBoxFragmentBuilder, and if an
object is *placed* by LayoutNG...
...LayoutBoxModelObject::RelativePositionOffset returns an empty offset.

This fixes a few small issues where previously we didn't resolve
percentages against the correct available size when contained within
an anonymous block.

Change-Id: I2c623b02cc5a4605d310596f5c60353a21236bff